value difference

+++ Sources/GroceryAppSharedModels/GroceryItemRequest.swift
@@ -9,10 +9,10 @@ import Foundation
 
 public struct GroceryItemRequest : Codable {
     public var title : String
-    public var price : Int
-    public var quantity : Double
+    public var price : Double
+    public var quantity : Int
     
-    public init(title: String, price: Int, quantity: Double) {
+    public init(title: String, price: Double, quantity: Int) {
         self.title = title
         self.price = price
         self.quantity = quantity
