+++ Sources/GroceryAppSharedModels/GroceryItemResponse.swift
@@ -10,10 +10,10 @@ import Foundation
 public struct GroceryItemResponse : Codable {
     public var id : UUID
     public var title : String
-    public var price : Int
-    public var quantity : Double
+    public var price : Double
+    public var quantity : Int
     
-   public init(id: UUID, title: String, price: Int, quantity: Double) {
+   public init(id: UUID, title: String, price: Double, quantity: Int) {
         self.id = id
         self.title = title
         self.price = price
